microservices/entity-extraction-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pytesseract==0.3.10
Pillow==10.1.0
PyMuPDF==1.23.8
opencv-python==4.6.0.66
numpy==1.24.3
python-dateutil==2.8.2
pdf2image==1.16.3
paddleocr==2.7.3
beautifulsoup4==4.12.2